+++ package.json
@@ -134,7 +134,7 @@
         },
         {
             "path": "./dist/minimal-component.js",
-            "maxSize": "7.5 kB"
+            "maxSize": "5.5 kB"
         }
     ]
 }
