+++ src/components/LazyMotion/index.tsx
@@ -1,9 +1,9 @@
-import { LazyProps } from "./types"
+// import { LazyProps } from "./types"
 
-export function LazyMotion({
-    renderer,
-    animation,
-    drag,
-    gestures,
-    layoutAnimation,
-}: LazyProps) {}
+// export function LazyMotion({
+//     renderer,
+//     animation,
+//     drag,
+//     gestures,
+//     layoutAnimation,
+// }: LazyProps) {}
