+++ src/events/utils.ts
@@ -1,4 +1,4 @@
-const isBrowser = typeof window !== "undefined"
+import { isBrowser } from "../utils/is-browser"
 
 // We check for event support via functions in case they've been mocked by a testing suite.
 export const supportsPointerEvents = () =>
