+++ src/motion/features/layout/index.ts
@@ -1,6 +1,8 @@
 import { FeatureBundle } from "../types"
 import { AnimateLayoutContextProvider } from "./Animate"
+import { MeasureContextProvider } from "./Measure"
 
 export const layoutAnimations: FeatureBundle = {
+    measureLayout: MeasureContextProvider,
     layoutAnimation: AnimateLayoutContextProvider,
 }
