--- src/motion/features/layout/use-snapshot-on-unmount.ts
@@ -1,28 +1,0 @@
-import { VisualElement } from "../../../render/types"
-import { useContext } from "react"
-import {
-    SharedLayoutContext,
-    FramerTreeLayoutContext,
-    isSharedLayout,
-} from "../../../context/SharedLayoutContext"
-import { useIsomorphicLayoutEffect } from "../../../utils/use-isomorphic-effect"
-
-export function useSnapshotOnUnmount(visualElement?: VisualElement) {
-    const syncLayout = useContext(SharedLayoutContext)
-    const framerSyncLayout = useContext(FramerTreeLayoutContext)
-
-    useIsomorphicLayoutEffect(
-        () => () => {
-            if (!visualElement) return
-
-            if (isSharedLayout(syncLayout)) {
-                syncLayout.remove(visualElement as any)
-            }
-
-            if (isSharedLayout(framerSyncLayout)) {
-                framerSyncLayout.remove(visualElement as any)
-            }
-        },
-        [visualElement]
-    )
-}
