+++ src/render/dom/motion.ts
@@ -4,6 +4,7 @@ import { createMotionProxy } from "./motion-proxy"
 import { createDomMotionConfig } from "./utils/create-config"
 import { gestures } from "../../motion/features/gestures"
 import { animations } from "../../motion/features/animations"
+import { drag } from "../../motion/features/drag"
 import { layoutAnimations } from "../../motion/features/layout"
 import { FeatureBundle } from "../../motion/features/types"
 import { createDomVisualElement } from "./create-visual-element"
@@ -11,6 +12,7 @@ import { createDomVisualElement } from "./create-visual-element"
 const featureBundle: FeatureBundle = {
     ...animations,
     ...gestures,
+    ...drag,
     ...layoutAnimations,
 }
