+++ src/render/dom/use-render.ts
@@ -1,3 +1,4 @@
+import * as React from "react"
 import { createElement } from "react"
 import { useHTMLProps } from "../html/use-props"
 import { filterProps } from "./utils/filter-props"
