+++ src/render/dom/value-types/get-as-type.ts
@@ -1,0 +1,10 @@
+import { ValueType } from "style-value-types"
+
+/**
+ * Provided a value and a ValueType, returns the value as that value type.
+ */
+export const getValueAsType = (value: any, type?: ValueType) => {
+    return type && typeof value === "number"
+        ? (type as any).transform(value)
+        : value
+}
