+++ src/render/dom/value-types/test.ts
@@ -1,0 +1,6 @@
+import { ValueType } from "style-value-types"
+
+/**
+ * Tests a provided value against a ValueType
+ */
+export const testValueType = (v: any) => (type: ValueType) => type.test(v)
