+++ src/render/dom/value-types/type-auto.ts
@@ -1,0 +1,9 @@
+import { ValueType } from "style-value-types"
+
+/**
+ * ValueType for "auto"
+ */
+export const auto: ValueType = {
+    test: (v: any) => v === "auto",
+    parse: (v) => v,
+}
