+++ src/render/dom/value-types/type-int.ts
@@ -1,0 +1,6 @@
+import { number } from "style-value-types"
+
+export const int = {
+    ...number,
+    transform: Math.round,
+}
