+++ src/render/dom/value-types/types.ts
@@ -1,0 +1,5 @@
+import { ValueType } from "style-value-types"
+
+export interface ValueTypeMap {
+    [key: string]: ValueType
+}
