+++ src/render/types.ts
@@ -1,4 +1,4 @@
-import { ComponentType } from "react"
+import * as React from "react"
 import { startAnimation } from "../animation/utils/transitions"
 import {
     Presence,
@@ -185,7 +185,7 @@ export type VisualElementOptions<Instance, RenderState = any> = {
 }
 
 export type CreateVisualElement<Instance> = (
-    Component: string | ComponentType,
+    Component: string | React.ComponentType,
     options: VisualElementOptions<Instance>
 ) => VisualElement<Instance>
