+++ src/render/utils/setters.ts
@@ -9,7 +9,8 @@ import {
 import { isNumericalString } from "../../utils/is-numerical-string"
 import { resolveFinalValueInKeyframes } from "../../utils/resolve-value"
 import { motionValue } from "../../value"
-import { findValueType, getAnimatableNone } from "../dom/utils/value-types"
+import { getAnimatableNone } from "../dom/value-types/animatable-none"
+import { findValueType } from "../dom/value-types/find"
 import { ResolvedValues, VisualElement } from "../types"
 import { AnimationDefinition } from "./animation"
 import { resolveVariant } from "./variants"
