+++ src/utils/is-browser.ts
@@ -1,0 +1,1 @@
+export const isBrowser = typeof window !== "undefined"
